inverter_model_maker/model_maker_web_v4/wsl_server/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
transformers>=4.50.0
torch>=2.3.0
accelerate>=0.33.0
bitsandbytes>=0.43.0
Pillow>=10.0.0
PyMuPDF>=1.24.0
autoawq
